Why use json when you can use python?
Instead of storing your config as a json object in a file, just use
python! Json is a pain to edit by hand, all those stupid braces and
random quotes. Python is much nicer. A little work has to be done in the
loading file to ensure all requisite values have reasonable defaults,
but that's not too hard.